Update 20220209 v2

+++ include/concrete/syscall.h
@@ -8,15 +8,15 @@ typedef unsigned long long u64;
 
 struct syscall_reg_t {
 	// arguments section
-	u64 rdi;  // rdi register
-	u64 rsi;  // rsi register
-	u64 rdx;  // rdx register
-	u64 rcx;  // rcx register
-	u64 r8;   // r8  register
-	u64 r9;   // r9  register
+	u32 rdi;  // rdi register
+	u32 rsi;  // rsi register
+	u32 rdx;  // rdx register
+	u32 rcx;  // rcx register
+	u32 r8;   // r8  register
+	u32 r9;   // r9  register
 
 	// retrun value section
-	u64 rax;  // rax register
+	u32 rax;  // rax register
 };
 
 void concrete_syscall(u16, struct syscall_reg_t*);
